requirements.txt
aadict==0.2.3
asset==0.6.13
Bootstrap-Flask==2.2.0
Bottleneck @ file:///opt/conda/conda-bld/bottleneck_1657175564434/work
certifi @ file:///croot/certifi_1671487769961/work/certifi
click==8.1.3
cssmin==0.2.0
cycler==0.11.0
Flask==2.2.2
Flask-Assets==2.0
Flask-SQLAlchemy==3.0.2
fonttools==4.38.0
globre==0.1.5
greenlet==2.0.1
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsmin==3.0.1
kiwisolver==1.4.4
lessc==0.1.3
lesscpy==0.15.1
MarkupSafe==2.1.1
matplotlib==3.5.3
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626179032232/work
mkl-service==2.4.0
numexpr @ file:///croot/numexpr_1668713893690/work
numpy @ file:///opt/conda/conda-bld/numpy_and_numpy_base_1653915516269/work
packaging @ file:///croot/packaging_1671697413597/work
pandas==1.3.5
Pillow==9.3.0
ply==3.11
psycopg2==2.9.5
pyparsing==3.0.9
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
python-dotenv==0.21.0
pytz @ file:///croot/pytz_1671697431263/work
rdkit==2022.9.3
six @ file:///tmp/build/80754af9/six_1644875935023/work
SQLAlchemy==1.4.46
typing_extensions==4.4.0
urllib3==1.26.14
webassets==2.0
Werkzeug==2.2.2
WTForms==3.0.1
zipp==3.11.0
